docs(schedule): document v1 usage and boundaries (#184)
