+++ packages/schedule/src/index.ts
@@ -1,22 +1,11 @@
 export { defineSchedule } from "./definition.ts"
-export type {} from "./registry-module.d.ts"
+export type {} from "./generated/registry.d.ts"
 export { ScheduleError } from "./errors.ts"
-export { createScheduleRun, executeRuntimeSchedule, executeSchedule, executeStaticSchedule } from "./runtime/execute.ts"
-export { schedules, validateRuntimeScheduleCron } from "./runtime/client.ts"
-export { createMemoryRuntimeScheduleStore, createMemoryScheduleRunStore } from "./runtime/store.ts"
-export {
-  getRuntimeScheduleStore,
-  getScheduleRunStore,
-  getScheduleRuntimeRegistry,
-  loadScheduleDefinition,
-  resetScheduleRuntime,
-  setRuntimeScheduleStore,
-  setScheduleRunStore,
-  setScheduleRuntimeRegistry,
-} from "./runtime/state.ts"
+export { setRuntimeScheduleStore, setScheduleRunStore, setScheduleRuntimeRegistry } from "./runtime/context.ts"
+export { createMemoryRuntimeScheduleStore, createMemoryScheduleRunStore } from "./runtime/memory-store.ts"
+export { schedules, validateRuntimeScheduleCron } from "./runtime/schedules.ts"
 
 export type {
-  DiscoveredScheduleDefinition,
   RuntimeScheduleCreateInput,
   RuntimeScheduleMetadata,
   RuntimeScheduleRecord,
@@ -26,7 +15,6 @@ export type {
   ScheduleRunAttemptStatus,
   ScheduleDefinition,
   ScheduleDefinitionOptions,
-  ScheduleDefinitionRegistry,
   ScheduleHandler,
   ScheduleRunContext,
   ScheduleRunError,
