+++ packages/schedule/src/nitro.ts
@@ -1,2 +1,2 @@
-export * from "./nitro/module.ts"
-export { default } from "./nitro/module.ts"
+export * from "./integrations/nitro/module.ts"
+export { default } from "./integrations/nitro/module.ts"
